Added authCookieOptions util
